WildcardMatch: Make the datapath port an Int

+++ midolman/src/main/java/org/midonet/sdn/flows/WildcardMatch.java
@@ -120,7 +120,7 @@ public class WildcardMatch implements Cloneable {
         return seenFields;
     }
 
-    private short inputPortNumber = 0;
+    private int inputPortNumber = 0;
     private long tunnelKey = 0L;
     private int tunnelSrc = 0;
     private int tunnelDst = 0;
@@ -255,7 +255,7 @@ public class WildcardMatch implements Cloneable {
     }
 
     @Nonnull
-    public WildcardMatch setInputPortNumber(short inputPortNumber) {
+    public WildcardMatch setInputPortNumber(int inputPortNumber) {
         usedFields.add(Field.InputPortNumber);
         this.inputPortNumber = inputPortNumber;
         return this;
@@ -269,7 +269,7 @@ public class WildcardMatch implements Cloneable {
     }
 
     @Nullable
-    public Short getInputPortNumber() {
+    public Integer getInputPortNumber() {
         return usedFields.contains(Field.InputPortNumber) ? inputPortNumber : null;
     }
 
@@ -1091,7 +1091,7 @@ public class WildcardMatch implements Cloneable {
 
                 case OpenVSwitch.FlowKey.Attr.InPort:
                     FlowKeyInPort inPort = as(flowKey, FlowKeyInPort.class);
-                    setInputPortNumber((short) inPort.getInPort());
+                    setInputPortNumber(inPort.getInPort());
                     break;
 
                 case OpenVSwitch.FlowKey.Attr.Ethernet:
